% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion Matrix}
\usage{
confusion_matrix(obs, pred, threshold = 0.5)
}
\arguments{
\item{obs}{a vector of observed values which must be 0 for absences and 1 for occurrences}

\item{pred}{a vector of the same length as \code{obs} representing the
predicted values. Values must be between 0 & 1 prepresenting a likelihood.}

\item{threshold}{a single threshold value between 0 & 1}
}
\value{
Returns a confusion matrix (table) of class 'confusion.matrix'
        representing counts of true & false presences and absences.
}
\description{
\code{confusion.matrix} calculates a confusion matrix. \cr
}
\note{
This function was adapted from `SDMtools` with a little change to provide new feature.
}

