% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluatePrecision}
\alias{evaluatePrecision}
\title{evaluatePrecision}
\usage{
evaluatePrecision(observed_value, predict_value, threshold)
}
\arguments{
\item{observed_value}{observed value vector or rasterLyaer, notice value of vector or rasterLayer must be 0 or 1.}

\item{predict_value}{predicted value vector or rasterLyaer, notice value of vector or rasterLayer must in range (0 ~ 1)}

\item{threshold}{threhold or cutoff translate predicted continue value into binary value (0 or 1),
normally the threshold value should in the range of (0 ~ 1), out of this range is also acceptable.}
}
\value{
a list of many citens, see details
}
\description{
evaluate the precision of prediction
}
\details{
Current return about 23 citens. them are :
         'acc', 'err', 'fpr', 'fall', 'tpr', 'rec', 'sens', 'fnr', 'miss',
         'tnr', 'spec', 'npv', 'pcmiss', 'rpp', 'rnp', 'rch', 'sar', 'auc',
         'rmse', 'kappa', 'tss'

         means about of those citens: \cr
         'acc': accuracy \cr
         'err': error rate \cr
         'fpr': false positive rate \cr
         'fall': fallout. same as fpr \cr
         'tpr': true positive rate \cr
         'rec': recall. same as tpr \cr
         'sens': sensitivity. same as tpr \cr
         'fnr': false negative rate \cr
         'miss': miss. same as fnr \cr
         'tnr': ture negative rate \cr
         'spec': specificity. same as tnr \cr
         'npv": negative predictive value \cr
         'pcmiss': prediction conditioned miss \cr
         'rpp': rate of positive predictions \cr
         'rnp': rate of positive predictions \cr
         'rch': ROC convex hull \cr
         'sar': score combinning performance measures of different characteristics \cr
         'auc': area under operator recevies curve \cr
         'kappa': Kappa statistic \cr
         'tss': true skill statistic or Hanssen-Kuiper score  \cr
}
\note{
Most power of this function are coming from package 'ROCR", we enchance a little
}

