% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimal_threshold.R
\name{optimalThreshold}
\alias{optimalThreshold}
\title{Find optimal threshold}
\usage{
optimalThreshold(observed_variable, predicted_variable)
}
\arguments{
\item{observed_variable}{Observed variable can be vector or rasterLyaer object}

\item{predicted_variable}{Predicted variable can be vector or rasterLyaer object}
}
\value{
a directory represent optimal threshold
}
\description{
Find optimal threshold
}
\details{
optimal threshold list: \cr
         'Default': threshold=0.5 \cr
         'Sens=Spec': sensitivity= specificity \cr
         'MaxSens+Spec': maximizes(sensitivity+specificity) \cr
         'MaxKappa': maximizes Kappa \cr
         'MaxPCC': maximizes PCC (precent correctly classified) \cr
         'PredPrev=Obs': predicted prevalence=observed prevalence \cr
         'ObsPrev': threshold=observed prevalence \cr
         'MeanProb': mean predicted probability \cr
         'MinROCdist': minizes distance between ROC plot and (0, 1) \cr
}
\note{
All of those optimal threshold are provide by `PresenceAbsence` package. This function just reformat it output
}

