\name{sdmutility-package}
\alias{sdmutility-package}
\alias{sdmutility}
\docType{package}
\title{
sdmutility
}
\description{
sdmutility is a utility (tool) package for SDM (Species Distribution Modelling)}
\details{
\tabular{ll}{
Package: \tab sdmutility\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-04-03\cr
License: \tab AGPL\cr
}
sdmutility is a part of SDM serial software
}
\author{
Xiaoquan Kong,
Renyan Duan,
Minyi Huang

Maintainer: Xiaoquan Kong <u1mail2me@gmail.com>
}

\keyword{ Species distribution modelling }
